+++ eds-aic-20230930/core/Properties/AssemblyInfo.cs
@@ -20,7 +20,7 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("5f4ce2d0-756f-11f1-8080-1deffc284f32")]
+[assembly: Guid("c3c882e0-7638-11f1-b162-613b02302d68")]
 
 // Version information for an assembly consists of the following four values:
 //
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("4.25.1.0")]
-[assembly: AssemblyFileVersion("4.26.0.0")]
+[assembly: AssemblyVersion("4.26.0.0")]
+[assembly: AssemblyFileVersion("4.26.1.0")]